+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - update README.md
 - fix bug on InnerThickOuterThin and InnerThickOuterThinCurve.
 - refactor to BoxDrawingBuilder class.
+- export BoxDrawingBuilder
 
 ## `0.0.7` 28 Dec 2022